Add Image Version to Docker-Compose to pull the relevant image (#2740)
This is to address 
- #2733 

Since a lot of times, the only problem is that docker image is not
consistent with the build. We have to somehow incorporate image version
into repository.

I don't insist to provide it this way. Maybe there are other automatic
ways which are better.

We can also calculate the relevant tag from git.